Move ApplicantSignup under SellerRegistration
