+++ src/bundles/SellerRegistration/ApplicantSignup.json
