+++ src/bundles/SellerRegistration/ApplicantSignupWidget.js
