+++ src/bundles/SellerRegistration/ApplicantSignupWidget.test.js
