+++ src/bundles/SellerRegistration/redux/helpers.js
