+++ src/bundles/SellerRegistration/redux/helpers.test.js
