+++ src/bundles/SellerRegistration/redux/modules/__tests__/application.test.js
