+++ src/bundles/SellerRegistration/redux/modules/reducer-signup.js
